Core 5002 add new member table (#1235)
* CORE-5002: add new member table

* CORE-5002: add link
