+++ templates/governance/supporters/index.jsx
@@ -1,8 +1,7 @@
-import React, { useState } from 'react'
-import { Button, Table, TextField } from '@oacore/design/lib/elements'
+import React from 'react'
 
 import styles from './supporters.module.scss'
-import imagePlaceholder from '../../../public/images/supporters/imagePlaceholder.svg'
+import MembershipTable from '../../../components/membership-table/membershipTable'
 
 import { Page, Markdown } from 'components'
 import { Layout, Hero, Section } from 'design-v2/components'
@@ -31,200 +30,61 @@ export const excludedIds = [
   3581, 197, 12800, 14335, 1249, 2313, 21853, 15201, 1012, 158,
 ]
 
-const GovernanceSupportersPageTemplate = ({ meta, supporters, members }) => {
-  const [searchValue, setSearchValue] = useState('')
-  const [displayedItems, setDisplayedItems] = useState(30)
-
-  const handleSearch = (event) => {
-    const { value } = event.target
-    setSearchValue(value)
-  }
-
-  const filteredMembers = members.filter((member) =>
-    member.organisation_name.toLowerCase().includes(searchValue.toLowerCase())
-  )
-
-  const sortedMembers = [...filteredMembers].sort((a, b) =>
-    a.organisation_name.localeCompare(b.organisation_name)
-  )
-
-  const renderBillingType = (member) => {
-    if (
-      excludedIds.includes(member.repo_id) ||
-      (Array.isArray(member.repo_id) &&
-        member.repo_id.some((id) => excludedIds.includes(id)))
-    ) {
-      return (
-        <a
-          target="_blank"
-          href="https://sparcopen.org/our-work/us-repository-network/discovery-pilot-project/"
-          rel="noreferrer"
-        >
-          USRN Partner
-        </a>
-      )
-    }
-
-    return (
-      <a target="_blank" href="https://core.ac.uk/membership" rel="noreferrer">
-        {member.billing_type} Member
-      </a>
-    )
-  }
-
-  const handleRedirect = (providerId) => {
-    const repoId = Array.isArray(providerId) ? providerId[0] : providerId
-    if (typeof window !== 'undefined')
-      window.location.href = `https://core.ac.uk/data-providers/${repoId}`
-  }
-
-  const getLink = (repoId) => {
-    if (Array.isArray(repoId)) {
-      // eslint-disable-next-line no-param-reassign
-      repoId = repoId.find((id) =>
-        fetch(`https://api.core.ac.uk/data-providers/${id}/logo`).then(() => id)
-      )
-    }
-
-    return (
-      <img
-        className={styles.repositoryLogo}
-        src={`https://api.core.ac.uk/data-providers/${repoId}/logo`}
-        onError={(e) => {
-          e.target.src = imagePlaceholder
-        }}
-        alt="logo"
+const GovernanceSupportersPageTemplate = ({ meta, supporters, members }) => (
+  <Page title={meta.title} description={meta.tagline}>
+    <Layout className={styles.container}>
+      <div className={styles.navWrapper}>
+        {supporters.items.headerLink.map((item) => (
+          <a className={styles.linkItem} href={item.href}>
+            {item.link}
+          </a>
+        ))}
+      </div>
+      <Hero
+        id={supporters.items.header.id}
+        image={supporters.items.header.image}
+        title={supporters.items.title}
+        description={supporters.items.description}
+        caption={supporters.items.tagline}
+        actions={supporters.items.actions}
+        reverse
+        spacing
       />
-    )
-  }
-
-  return (
-    <Page title={meta.title} description={meta.tagline}>
-      <Layout className={styles.container}>
-        <div className={styles.navWrapper}>
-          {supporters.items.headerLink.map((item) => (
-            <a className={styles.linkItem} href={item.href}>
-              {item.link}
-            </a>
+      <Section id="our-principles" className={styles.ourPrinciples}>
+        <h4>{supporters.items.ourPrinciples.title}</h4>
+        <div className={styles.principlesWrapper}>
+          {supporters.items.ourPrinciples.cardsDescription.map((plan) => (
+            <CardDescription key={plan.title} plan={plan} />
           ))}
         </div>
-        <Hero
-          id={supporters.items.header.id}
-          image={supporters.items.header.image}
-          title={supporters.items.title}
-          description={supporters.items.description}
-          caption={supporters.items.tagline}
-          actions={supporters.items.actions}
-          reverse
-          spacing
-        />
-        <Section id="our-principles" className={styles.ourPrinciples}>
-          <h4>{supporters.items.ourPrinciples.title}</h4>
-          <div className={styles.principlesWrapper}>
-            {supporters.items.ourPrinciples.cardsDescription.map((plan) => (
-              <CardDescription key={plan.title} plan={plan} />
-            ))}
-          </div>
-        </Section>
-        <Section id="principles" className={styles.principles}>
-          <div className={styles.imageWrapper}>
-            <img
-              className={styles.image}
-              src={supporters.items.principles.image}
-              alt="support"
-            />
-          </div>
-          <article className={styles.content}>
-            <Markdown>{supporters.items.principles.description}</Markdown>
-          </article>
-        </Section>
-        <Section id="how-it-works" className={styles.howItWorks}>
-          <h4>{supporters.items.howItWorks.title}</h4>
-          <div className={styles.howItWorksWrapper}>
-            {supporters.items.howItWorks.services.map((plan) => (
-              <Card key={plan.title} plan={plan} />
-            ))}
-          </div>
-        </Section>
-        <Section
-          id={supporters.items.supporters.id}
-          className={styles.supporters}
-        >
-          <h3>{supporters.items.supporters.title}</h3>
-          <form>
-            <TextField
-              className={styles.search}
-              value={searchValue}
-              onChange={handleSearch}
-              id="search"
-              type="search"
-              name="institution"
-              label="Find Organisation"
-              placeholder="e.g Open University"
-            />
-          </form>
-          <Table>
-            <Table.Head>
-              <Table.Row>
-                <Table.HeadCell>Organisation name</Table.HeadCell>
-                <Table.HeadCell>Country</Table.HeadCell>
-                <Table.HeadCell>Membership status</Table.HeadCell>
-              </Table.Row>
-            </Table.Head>
-            <Table.Body>
-              {sortedMembers.slice(0, displayedItems).map((member) => (
-                <Table.Row
-                  className={styles.tableRow}
-                  key={member.organisation_id}
-                >
-                  <div className={styles.organisationWrapper}>
-                    <div className={styles.logoWrapper}>
-                      {getLink(member.repo_id)}
-                    </div>
-                    <Table.Cell onClick={() => handleRedirect(member.repo_id)}>
-                      {member.organisation_name}
-                    </Table.Cell>
-                  </div>
-                  <Table.Cell>{member.country}</Table.Cell>
-                  <Table.Cell className={styles.memberColumn}>
-                    {renderBillingType(member)}
-                  </Table.Cell>
-                </Table.Row>
-              ))}
-              {filteredMembers.length === 0 && (
-                <Table.Row>
-                  <Table.Cell colSpan={2}>
-                    No results. Please update your query
-                  </Table.Cell>
-                </Table.Row>
-              )}
-            </Table.Body>
-            <Table.Footer>
-              <Table.Row className={styles.paginationRow}>
-                <Table.Cell colSpan={3}>
-                  <p className={styles.paginationText}>
-                    Showing 1 -{' '}
-                    {Math.min(displayedItems, filteredMembers.length)}
-                  </p>
-                  <Button
-                    onClick={() =>
-                      setDisplayedItems((prevCount) =>
-                        Math.min(prevCount + 10, filteredMembers.length)
-                      )
-                    }
-                    variant="outlined"
-                    disabled={displayedItems >= filteredMembers.length}
-                  >
-                    Show More
-                  </Button>
-                </Table.Cell>
-              </Table.Row>
-            </Table.Footer>
-          </Table>
-        </Section>
-      </Layout>
-    </Page>
-  )
-}
+      </Section>
+      <Section id="principles" className={styles.principles}>
+        <div className={styles.imageWrapper}>
+          <img
+            className={styles.image}
+            src={supporters.items.principles.image}
+            alt="support"
+          />
+        </div>
+        <article className={styles.content}>
+          <Markdown>{supporters.items.principles.description}</Markdown>
+        </article>
+      </Section>
+      <Section id="how-it-works" className={styles.howItWorks}>
+        <h4>{supporters.items.howItWorks.title}</h4>
+        <div className={styles.howItWorksWrapper}>
+          {supporters.items.howItWorks.services.map((plan) => (
+            <Card key={plan.title} plan={plan} />
+          ))}
+        </div>
+      </Section>
+      <MembershipTable
+        title={supporters.items.supporters.title}
+        members={members}
+        excludedIds={excludedIds}
+      />
+    </Layout>
+  </Page>
+)
 
 export default GovernanceSupportersPageTemplate
